reg save

+++ stm32/jt/TouchGFX/gui/include/gui/mainscreen_screen/mainScreenPresenter.hpp
@@ -29,7 +29,7 @@ public:
 
     uint16_t getWeldingTimeSec();
 
-    uint32_t getWeldingAmps();
+    float getWeldingAmps();
 
 private:
     mainScreenPresenter();
